Fix 1.0.0.5

+++ CS/ConnectServer/QueueHandle.h
@@ -9,7 +9,7 @@
 struct QUEUE_INFO
 {
 	WORD index;
-	DWORD size;
+	WORD size;
 	BYTE head;
 	BYTE buff[MAX_BUFFER_QUEUE_SIZE];
 };
